+++ R/layer_to_vector.R
@@ -2,7 +2,8 @@
 #'
 #' Extracts the values of a single layer in a \code{stars} object to a vector. Cell values are ordered from top-left corner to the right.
 #'
-#' @param	r	A raster (class \code{stars}) with two dimensions: \code{x} and \code{y}, i.e., a single-band raster
+#' @param	x	A raster (class \code{stars}) with two dimensions: \code{x} and \code{y}, i.e., a single-band raster.
+#' @param check Whether to check (and fix if necessary) that input has one attribute, one layer and x-y as dimensions 1-2 (default is \code{TRUE}).
 #' @return A vector with cell values, ordered by rows, starting from the top left corner (north-west) and to the right. 
 #'
 #' @examples
@@ -12,15 +13,17 @@
 #'
 #' @export
 
-layer_to_vector = function(r) {
+layer_to_vector = function(x, check = TRUE) {
 
   # Checks
-  r = check_one_attribute(r)
-  r = check_spatial_dimensions(r)
-  r = check_one_layer(r)
+  if(check) {
+    x = check_one_attribute(x)
+    x = check_spatial_dimensions(x)
+    x = check_one_layer(x)
+  }
 
   # To matrix
-  m = as.vector(r[[1]])
+  m = as.vector(x[[1]])
 
   # Return
   return(m)